requirements.txt
streamlit>=1.28.0
docker>=6.1.0
requests>=2.31.0
hypothesis>=6.88.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.7.0